+++ tests/test_visitors/test_ast/test_naming/conftest.py
@@ -1,6 +1,7 @@
 import pytest
 
-from wemake_python_styleguide.compat.constants import PY38
+from wemake_python_styleguide.compat.constants import PY38, PY310
+from wemake_python_styleguide.constants import UNUSED_PLACEHOLDER
 
 # Imports:
 
@@ -161,6 +162,27 @@ except Exception as {0}:
     raise
 """
 
+# Pattern matching:
+
+match_variable = """
+match some_value:
+    case {0}:
+        ...
+"""
+
+match_inner = """
+match some_value:
+    case [{0}]:
+        ...
+"""
+
+# This is the only case where we don't allow unused variables.
+match_as_explicit = """
+match some_value:
+    case [] as {0}:
+        ...
+"""
+
 
 # Fixtures:
 
@@ -220,6 +242,12 @@ if PY38:
         lambda_posonly_argument,
         assignment_expression,
     }
+if PY310:
+    _ALL_FIXTURES |= {
+        match_variable,
+        match_as_explicit,
+        match_inner,
+    }
 
 _FOREIGN_NAMING_PATTERNS = frozenset((
     foreign_attribute,
@@ -260,6 +288,10 @@ _FORBIDDEN_BOTH_RAW_AND_PROTECTED_UNUSED = frozenset((
 if PY38:
     _FORBIDDEN_BOTH_RAW_AND_PROTECTED_UNUSED |= {
         assignment_expression,
+    }
+if PY310:
+    _FORBIDDEN_BOTH_RAW_AND_PROTECTED_UNUSED |= {
+        match_as_explicit,
     }
 
 _FORBIDDEN_RAW_UNUSED = _FORBIDDEN_BOTH_RAW_AND_PROTECTED_UNUSED | {
@@ -331,3 +363,12 @@ def forbidden_protected_unused_template(request):
 def allowed_protected_unused_template(request):
     """Returns template that allows defining protected unused variables."""
     return request.param
+
+
+@pytest.fixture()
+def skip_match_case_syntax_error():
+    """Returns a helper that skips tests when `_` is used with pattern match."""
+    def factory(template: str, var_name: str) -> None:
+        if var_name == UNUSED_PLACEHOLDER and template == match_as_explicit:
+            pytest.skip('"_" cannot be used as "case" target')
+    return factory